+++ pkg/okteto/preview_test.go
@@ -16,7 +16,6 @@ package okteto
 import (
 	"context"
 	"errors"
-	"fmt"
 	"strings"
 	"testing"
 
@@ -175,7 +174,8 @@ func TestDeployPreview(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			pc := previewClient{
-				client: tc.input.client,
+				client:             tc.input.client,
+				namespaceValidator: newNamespaceValidator(),
 			}
 			response, err := pc.DeployPreview(context.Background(), tc.input.name, "", "", "", "", "", tc.input.variables)
 			assert.ErrorIs(t, err, tc.expected.err)
@@ -691,61 +691,6 @@ func TestTranslatePreviewErr(t *testing.T) {
 			pc := previewClient{}
 			err := pc.translateErr(tc.input.err, tc.input.name)
 			assert.ErrorIs(t, err, tc.expected.err)
-		})
-	}
-}
-
-func Test_validateNamespaceName(t *testing.T) {
-	tests := []struct {
-		name          string
-		namespace     string
-		expectedError bool
-		errorMessage  string
-	}{
-		{
-			name:          "ok-namespace-starts-with-letter",
-			namespace:     "argo-yournamespace",
-			expectedError: false,
-			errorMessage:  "",
-		},
-		{
-			name:          "ok-namespace-starts-with-number",
-			namespace:     "1-argo-yournamespace",
-			expectedError: false,
-			errorMessage:  "",
-		},
-		{
-			name:          "wrong-namespace-starts-with-unsupported-character",
-			namespace:     "-argo-yournamespace",
-			expectedError: true,
-			errorMessage:  "Malformed namespace name",
-		},
-		{
-			name:          "wrong-namespace-unsupported-character",
-			namespace:     "argo/test-yournamespace",
-			expectedError: true,
-			errorMessage:  "Malformed namespace name",
-		},
-		{
-			name:          "wrong-namespace-exceeded-char-limit",
-			namespace:     fmt.Sprintf("%s-yournamespace", strings.Repeat("test", 20)),
-			expectedError: true,
-			errorMessage:  "Exceeded number of character",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := validateNamespace(tt.namespace, "namespace")
-			if err != nil && !tt.expectedError {
-				t.Errorf("Expected error but no error found")
-			}
-			if err != nil && !tt.expectedError {
-				if !strings.Contains(err.Error(), tt.errorMessage) {
-					t.Errorf("Expected %s, but got %s", tt.errorMessage, err.Error())
-				}
-			}
-
 		})
 	}
 }
